Web Scraping Real Estate Listings for Rental
This project is focused on automating the extraction of real estate listings for rental purposes. It leverages undetected_chromedriver and BeautifulSoup for navigating and parsing the Idealista website. The primary goal is to facilitate a study on market trends and rental prices in these areas by collecting and analyzing data systematically.

**Key Features:**
- Automated navigation using Selenium with undetected_chromedriver to interact with web pages while minimizing detection risks.
- Data parsing with BeautifulSoup to efficiently extract listing details such as title, price, description, and advertiser information.
- Detailed logging for monitoring the scraping process and identifying potential issues.
- Output of scraped data in both JSON and Excel formats for versatility and accessibility.
- Implementation of respectful delay intervals between requests to reduce the load on the website's servers.
- Dynamic tracking of listing statuses (new, available, rented) based on their presence in subsequent scraping sessions.

**Ethical Considerations and Study Intentions:**
This project is developed with the intention of conducting a personal study on real estate market trends and rental pricing within specified regions. It is not intended for commercial use or any application that would conflict with the terms of service of the Idealista website or any other source. All scraping activities are designed to be minimally invasive and comply with the website's robots.txt rules, reflecting a commitment to ethical data collection practices. This project serves as an educational tool and a means to contribute to the understanding of real estate dynamics in the Alvalade areas.

Please note that this script is for educational and informational purposes only. Users are responsible for ensuring their use of the script complies with all relevant laws and website terms of service.